fix(core): a decimal operand is named 'decimal' in every format and subscript message; python's Decimal grammar recorded as a deviation (#2014)
`PyFormat.PyTypeName` Float arm: `System.Decimal` → "decimal"; `double`,
`float32` and `Half` stay "float", every integer width stays "int" (they
have python twins under those names). Rung 2 — the ONE name table in Core;
the compiler's static twin reads it through `FormatOperandTypeName`, so the
static SPY0609, the runtime ValueError and the str.format subscript
TypeError all move together. Sharpy's `decimal` keeps the float grammar
(Axiom 1); python's `decimal.Decimal.__format__` grammar is recorded as the
new `decimal-format-grammar` deviation (docs/deviations.yaml; python3
3.12.13: `format(Decimal('1.5'), 'e')` = '1.5e+0', `'d'`/`'s'`/`'#c'` =
`invalid format string`, subscript names 'decimal.Decimal'; Sharpy:
'1.500000e+00'), oracle ledger regenerated (+8, 88 deviations), and one
fstrings.md paragraph under Refusal Precedence.

Direction record (base = sharpyc @ acd1d40a2, laneC-base worktree → this
commit), `m: decimal = decimal(1.5)`:
- `format(m, spec)` spec in a variable, d / s / #c: ValueError "Unknown
  format code 'd'|'s'|'c' for object of type 'float'" → "… type 'decimal'".
- literal `format(m, "d")`: SPY0609 "… type 'float'" → "… type 'decimal'".
- `"{0[0]}".format(m)`: TypeError "'float' object is not subscriptable" →
  "'decimal' object is not subscriptable".
- `.2f` → '1.50' unchanged (control); float32 'd' → 'float', int8 's' →
  'int', float32 subscript → 'float' unchanged (controls).
- message text moves only; no program newly refused or newly accepted.

Moved expectation (the decision itself, Decision 15):
`PyFormatTests.TypeNameCells` `typeof(decimal)`: "float" → "decimal".

New rows: `FormatOperandKindParityTests.FormatRefusal_NamesTheOperandType_StaticAndRuntime`
({decimal} x {d, s, #c, .2f control} + float32/int8 controls, each asserting
the runtime text AND the static SPY0609 message) and
`FieldSubscript_NamesTheOperandType` (decimal, float32, int8).

Mutation (worktree sharpy.worktrees/laneD-mut, restored with cp): revert
the arm to `return "float";` → red (FormatOperandKindParity 4 failed: the
decimal d/s/#c rows + the decimal subscript row; Core PyFormatTests 1
failed: the decimal name cell); restored → green (16 passed; Core.Tests
3865/0/0). Compiler format filter (FormatOperandKindParity,
FormatEngineConsumerParity, FormatSpecStaticTwinRoute,
FStringReplacementFieldMatrix) 32/0/0 measured in the same worktree.

+++ src/Sharpy.Core.Tests/PyFormatTests.cs
@@ -738,7 +738,7 @@ public class PyFormatTests
         yield return new object[] { typeof(bool), "bool" };
         yield return new object[] { typeof(double), "float" };
         yield return new object[] { typeof(float), "float" };
-        yield return new object[] { typeof(decimal), "float" };
+        yield return new object[] { typeof(decimal), "decimal" };   // Sharpy's own name: python's Decimal has another grammar (#2014)
         foreach (var t in new[] { typeof(int), typeof(long), typeof(short), typeof(byte), typeof(sbyte), typeof(uint), typeof(ulong), typeof(ushort) })
         {
             yield return new object[] { t, "int" };
